+++ code/lib/builder-vite/src/index.ts
@@ -14,6 +14,7 @@ import { createViteServer } from './vite-server';
 import { build as viteBuild } from './build';
 import type { ExtendedOptions } from './types';
 
+export { withoutVitePlugins } from './utils/without-vite-plugins';
 export type { TypescriptOptions } from '@storybook/core-common';
 
 // Storybook's Stats are optional Webpack related property