+++ code/lib/client-api/src/StoryStoreFacade.ts
@@ -15,7 +15,6 @@ import type {
   IndexEntry,
 } from '@storybook/store';
 import { logger } from '@storybook/client-logger';
-import deprecate from 'util-deprecate';
 import type { DocsOptions } from '@storybook/core-common';
 
 export interface GetStorybookStory<TFramework extends AnyFramework> {
@@ -28,9 +27,6 @@ export interface GetStorybookKind<TFramework extends AnyFramework> {
   fileName: string;
   stories: GetStorybookStory<TFramework>[];
 }
-
-const docs2Warning = deprecate(() => {},
-`You cannot use \`.mdx\` files without using \`storyStoreV7\`. Consider upgrading to the new store.`);
 
 export class StoryStoreFacade<TFramework extends AnyFramework> {
   projectAnnotations: NormalizedProjectAnnotations<TFramework>;
@@ -146,7 +142,6 @@ export class StoryStoreFacade<TFramework extends AnyFramework> {
   // NOTE: we could potentially share some of this code with the stories.json generation
   addStoriesFromExports(fileName: Path, fileExports: ModuleExports) {
     if (fileName.match(/\.mdx$/) && !fileName.match(/\.stories\.mdx$/)) {
-      docs2Warning();
       return;
     }