+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+## 7.0.0-alpha.41 (October 21, 2022)
+
+#### Features
+
+-   Add `@storybook/nextjs` framework [#19382](https://github.com/storybooks/storybook/pull/19382)
+-   CLI: Enable `@storybook/nextjs` framework [#19478](https://github.com/storybooks/storybook/pull/19478)
+-   CLI: Automigrate from MDX1 to MDX2 [#19568](https://github.com/storybooks/storybook/pull/19568)
+
+#### Maintenance
+
+-   Remove warning of removed feature in lib/client-api [#19544](https://github.com/storybooks/storybook/pull/19544)
+-   Remove vite-plugin-svelte-kit when detected [#19522](https://github.com/storybooks/storybook/pull/19522)
+
+#### Build
+
+-   Delete Svelte example [#19549](https://github.com/storybooks/storybook/pull/19549)
+-   Fix circle test results [#19552](https://github.com/storybooks/storybook/pull/19552)
+
+#### Dependency Upgrades
+
+-   Update the version of the "update-notifier" package [#19569](https://github.com/storybooks/storybook/pull/19569)
+
+## 7.0.0-alpha.40 (October 20, 2022)
+
+#### Breaking Changes
+
+- Addon-docs: Upgrade to MDXv2 [#19495](https://github.com/storybooks/storybook/pull/19495)
+
+#### Bug Fixes
+
+- Addon-docs: Don't generate docs page entries for CSF files with no stories [#19529](https://github.com/storybooks/storybook/pull/19529)
+
+#### Maintenance
+
+- Remove deprecate features from preview-web [#19540](https://github.com/storybooks/storybook/pull/19540)
+- Remove deprecated features in lib/api [#19539](https://github.com/storybooks/storybook/pull/19539)
+- Remove default selection in docblocks [#19537](https://github.com/storybooks/storybook/pull/19537)
+
+#### Build
+
+- Remove .git folder when generating repros [#19535](https://github.com/storybooks/storybook/pull/19535)
+- Some task runner tweaks + move test-runner into sandbox task. [#19505](https://github.com/storybooks/storybook/pull/19505)
+- Build builder-webpack5 with ts-up [#19435](https://github.com/storybooks/storybook/pull/19435)
+
 ## 7.0.0-alpha.39 (October 19, 2022)
 
 #### Breaking Changes